+++ src/com_server/api_server.py
@@ -14,6 +14,7 @@ import waitress
 from flask_cors import CORS
 
 from . import base_connection, connection  # for typing
+from . import disconnect
 
 
 class EndpointExistsException(Exception):
@@ -74,7 +75,7 @@ class RestApiHandler:
         has_register_recall: bool = True,
         add_cors: bool = False,
         catch_all_404s: bool = True,
-        **kwargs,
+        **kwargs: t.Dict[str, t.Any],
     ) -> None:
         """Constructor for class
 
@@ -104,8 +105,8 @@ class RestApiHandler:
 
         # other
         self._all_endpoints = (
-            []
-        )  # list of all endpoints in tuple (endpoint str, resource class)
+            []  # list of all endpoints in tuple (endpoint str, resource class)
+        )
         self._registered = (
             None  # keeps track of who is registered; None if not registered
         )
@@ -188,7 +189,9 @@ class RestApiHandler:
             resource.conn = self._conn
 
             # req methods; _self is needed as these will be part of class functions
-            def _get(_self, *args, **kwargs):
+            def _get(
+                _self, *args: t.Tuple[t.Any], **kwargs: t.Dict[str, t.Any]
+            ) -> t.Any:
                 ip = flask.request.remote_addr
                 if self._has_register_recall and (
                     not self._registered or self._registered != ip
@@ -202,7 +205,9 @@ class RestApiHandler:
                         _self, *args, **kwargs
                     )  # resource.[METHOD]() will be replaced with resource._[METHOD] below
 
-            def _post(_self, *args, **kwargs):
+            def _post(
+                _self, *args: t.Tuple[t.Any], **kwargs: t.Dict[str, t.Any]
+            ) -> t.Any:
                 ip = flask.request.remote_addr
                 if self._has_register_recall and (
                     not self._registered or self._registered != ip
@@ -214,7 +219,9 @@ class RestApiHandler:
                 else:
                     return resource._post(_self, *args, **kwargs)
 
-            def _head(_self, *args, **kwargs):
+            def _head(
+                _self, *args: t.Tuple[t.Any], **kwargs: t.Dict[str, t.Any]
+            ) -> t.Any:
                 ip = flask.request.remote_addr
                 if self._has_register_recall and (
                     not self._registered or self._registered != ip
@@ -226,7 +233,9 @@ class RestApiHandler:
                 else:
                     return resource._head(_self, *args, **kwargs)
 
-            def _put(_self, *args, **kwargs):
+            def _put(
+                _self, *args: t.Tuple[t.Any], **kwargs: t.Dict[str, t.Any]
+            ) -> t.Any:
                 ip = flask.request.remote_addr
                 if self._has_register_recall and (
                     not self._registered or self._registered != ip
@@ -238,7 +247,9 @@ class RestApiHandler:
                 else:
                     return resource._put(_self, *args, **kwargs)
 
-            def _delete(_self, *args, **kwargs):
+            def _delete(
+                _self, *args: t.Tuple[t.Any], **kwargs: t.Dict[str, t.Any]
+            ) -> t.Any:
                 ip = flask.request.remote_addr
                 if self._has_register_recall and (
                     not self._registered or self._registered != ip
@@ -273,7 +284,7 @@ class RestApiHandler:
 
         return _outer
 
-    def add_resource(self, *args, **kwargs) -> None:
+    def add_resource(self, *args: t.Tuple[t.Any], **kwargs: t.Dict[str, t.Any]) -> None:
         """Calls `flask_restful.add_resource`. Allows adding endpoints
         without needing a connection.
 
@@ -284,7 +295,7 @@ class RestApiHandler:
 
         return self._api.add_resource(*args, **kwargs)
 
-    def run_dev(self, **kwargs) -> None:
+    def run_dev(self, **kwargs: t.Dict[str, t.Any]) -> None:
         """Launches the Flask app as a development server.
 
         All arguments in `**kwargs` will be passed to `Flask.run()`.
@@ -307,11 +318,15 @@ class RestApiHandler:
         for endpoint, resource in self._all_endpoints:
             self._api.add_resource(resource, endpoint)
 
+        # add disconnect handler, verbose is True
+        _disconnect_handler = disconnect.Reconnector(self._conn, True)
+        _disconnect_handler.start()
+
         self._app.run(**kwargs)
 
         self._conn.disconnect()  # disconnect if stop running
 
-    def run_prod(self, **kwargs) -> None:
+    def run_prod(self, **kwargs: t.Dict[str, t.Any]) -> None:
         """Launches the Flask app as a Waitress production server.
 
         All arguments in `**kwargs` will be passed to `waitress.serve()`.
@@ -327,6 +342,10 @@ class RestApiHandler:
         # register all endpoints to flask_restful
         for endpoint, resource in self._all_endpoints:
             self._api.add_resource(resource, endpoint)
+
+        # add disconnect handler, verbose is False
+        _disconnect_handler = disconnect.Reconnector(self._conn, False)
+        _disconnect_handler.start()
 
         waitress.serve(self._app, **kwargs)