+++ CHANGELOG.md
@@ -14,6 +14,8 @@ Previous version: 0.1b0
 - Added new endpoint `/connection_state`, getting some properties of the `Connection` object, addressing [#60](https://github.com/jonyboi396825/COM-Server/issues/60)
 - Added option to remove 0.01 second delay at end of IO thread, addressing [#68](https://github.com/jonyboi396825/COM-Server/issues/68)
 - `exception` in `BaseConnection` and `Connection` objects is now **DEPRECATED**
+- Added disconnect handling in server (reconnects automatically, responds with `500` when disconnected)
+- Cleaned up typing in source code
 
 # 0.1 Beta Release 0